+++ tests/Multitarget.Tests/Main.fs
@@ -416,6 +416,7 @@ let all =
         nestedTestCase
         equalityTestCases
         pendingListTestCases
+        ParityTests.all
         //focusedTestCases
         //errorTestCases
         //failedTestCases